+++ src/components/Breadcrumbs/BreadcrumbItem.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-	import { El } from '../Base'
+	import { El } from '$lib/components'
+
 	import type { BreadcrumbItemProps } from './Breadcrumb.types'
 
 	type $$Props = BreadcrumbItemProps
