+++ src/components/Footer/Footer.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-	import { El } from '../Base'
+	import { El } from '$lib/components'
+
 	import type { FooterProps } from './Footer.types'
 
 	type $$Props = FooterProps
