+++ src/components/Input/Input.types.ts
@@ -1,4 +1,4 @@
-import type { ElProps } from '../Base'
+import type { ElProps } from '$lib/components'
 
 export const INPUT_SIZES = ['sm', 'md', 'lg'] as const
 export type InputSizes = typeof INPUT_SIZES[number]
