+++ src/components/Spinner/Spinner.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-	import { El } from '../Base'
+	import { El } from '$lib/components'
+
 	import type { SpinnerProps } from './Spinner.types'
 
 	type $$Props = SpinnerProps
