+++ src/components/Table/TableBody.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-	import { El } from '../Base'
+	import { El } from '$lib/components'
+
 	import type { TableBodyProps } from './Table.types'
 
 	type $$Props = TableBodyProps
