+++ src/components/Table/TableCell.svelte
@@ -1,11 +1,13 @@
 <script lang="ts">
-  import { getContext } from 'svelte'
-	import { El } from '../Base'
+	import { getContext } from 'svelte'
+
+	import { El } from '$lib/components'
+
 	import type { TableCellProps } from './Table.types'
 
 	type $$Props = TableCellProps
 
-	export let tag: $$Props['tag'] = 'td';
+	export let tag: $$Props['tag'] = 'td'
 	export let cssPrefix: $$Props['cssPrefix'] = 'table-cell'
 
 	export let color: $$Props['color'] = undefined
@@ -18,7 +20,7 @@
 	$: cssProps = {
 		color,
 		truncate,
-		active
+		active,
 	}
 
 	$: {
