+++ src/components/Toast/Toast.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-	import { El } from '../Base'
+	import { El } from '$lib/components'
+
 	import type { ToastProps } from './Toast.types'
 
 	type $$Props = ToastProps
