+++ src/components/Toast/Toast.types.ts
@@ -1,15 +1,15 @@
-import type { ElProps } from '../Base'
+import type { ElProps } from '$lib/components'
 
 export interface ToastProps extends Partial<ElProps> {
 	show?: boolean
 }
 
 export interface ToastBodyProps extends Partial<ElProps> {
-    // 
+	//
 }
 
 export interface ToastContainerProps extends Partial<ElProps> {
-    // 
+	//
 }
 
 export interface ToastHeaderProps extends Partial<ElProps> {
