+++ src/components/index.ts
@@ -1,10 +1,8 @@
-export * from './Base'
 export * from './Accordion'
 export * from './Alert'
 export * from './App'
 export * from './Avatar'
 export * from './Badge'
-export * from './Base'
 export * from './Breadcrumbs'
 export * from './Button'
 export * from './Card'
@@ -14,6 +12,7 @@ export * from './Confirm'
 export * from './Dialog'
 export * from './Divider'
 export * from './Dot'
+export * from './El'
 export * from './FileUploader'
 export * from './Form'
 export * from './Footer'
